+++ packages/ui/src/components/link-button/index.tsx
@@ -1,0 +1,7 @@
+export type {
+  TRLinkButtonAppearance,
+  TRLinkButtonIntent,
+  TRLinkButtonProps,
+  TRLinkButtonUiSize,
+} from './link-button.js';
+export { TRLinkButton } from './link-button.js';